Move pkg/encoding to root
