+++ clients/tests/retrieval_client_test.go
@@ -13,8 +13,8 @@ import (
 	"github.com/Layr-Labs/eigenda/core/encoding"
 	coreindexer "github.com/Layr-Labs/eigenda/core/indexer"
 	coremock "github.com/Layr-Labs/eigenda/core/mock"
+	kzgEncoder "github.com/Layr-Labs/eigenda/encoding/kzg"
 	indexermock "github.com/Layr-Labs/eigenda/indexer/mock"
-	"github.com/Layr-Labs/eigenda/pkg/encoding/kzgEncoder"
 	"github.com/consensys/gnark-crypto/ecc/bn254"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
