+++ encoding/kzg/batchCommitEquivalence.go
