+++ encoding/kzg/encoder_fuzz_test.go
@@ -1,9 +1,10 @@
-package encoder_test
+package kzgEncoder_test
 
 import (
 	"testing"
 
-	rs "github.com/Layr-Labs/eigenda/pkg/encoding/encoder"
+	kzgRs "github.com/Layr-Labs/eigenda/encoding/kzg"
+	rs "github.com/Layr-Labs/eigenda/encoding/rs"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -12,14 +13,21 @@ func FuzzOnlySystematic(f *testing.F) {
 	f.Add(GETTYSBURG_ADDRESS_BYTES)
 	f.Fuzz(func(t *testing.T, input []byte) {
 
+		group, _ := kzgRs.NewKzgEncoderGroup(kzgConfig, true)
+
 		params := rs.GetEncodingParams(10, 3, uint64(len(input)))
-		enc, err := rs.NewEncoder(params, true)
+		enc, err := group.NewKzgEncoder(params)
 		if err != nil {
 			t.Errorf("Error making rs: %q", err)
 		}
 
 		//encode the data
-		_, frames, _, err := enc.EncodeBytes(input)
+		_, _, _, frames, _, err := enc.EncodeBytes(input)
+
+		for _, frame := range frames {
+			assert.NotEqual(t, len(frame.Coeffs), 0)
+		}
+
 		if err != nil {
 			t.Errorf("Error Encoding:\n Data:\n %q \n Err: %q", input, err)
 		}
