+++ encoding/rs/decode.go
