+++ encoding/rs/encode_test.go
@@ -7,7 +7,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	rs "github.com/Layr-Labs/eigenda/pkg/encoding/encoder"
+	rs "github.com/Layr-Labs/eigenda/encoding/rs"
 )
 
 func TestEncodeDecode_InvertsWhenSamplingAllFrames(t *testing.T) {
