+++ encoding/rs/encoder.go
