+++ encoding/rs/utils.go
@@ -4,7 +4,7 @@ import (
 	"errors"
 	"math"
 
-	rb "github.com/Layr-Labs/eigenda/pkg/encoding/utils/reverseBits"
+	rb "github.com/Layr-Labs/eigenda/encoding/utils/reverseBits"
 
 	bls "github.com/Layr-Labs/eigenda/pkg/kzg/bn254"
 )
