+++ encoding/utils/toeplitz/cir.go
