+++ encoding/utils/toeplitz/toeplitz.go
